requirements.txt
# Core AI
anthropic>=0.40.0
sentence-transformers>=2.3.0
chromadb>=0.4.22
tiktoken>=0.5.0

# Hybrid search
rank-bm25>=0.2.2

# Multimodal - PDF & Images
PyMuPDF>=1.24.0
Pillow>=10.2.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Auth
bcrypt>=4.1.0
PyJWT>=2.8.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
rich>=13.7.0
tenacity>=8.2.0

# PageIndex-style local engine (no external dependencies — uses Claude + PyMuPDF)
